Calibrates input images
ok so here's the thing
Using blender generated images, which doesn't feature lens distortion, the calibration bends straight lines into convex shapes which is, on one hand, really nice, but on the other hand, kinda wrong.
It is my belief that it only happens because Blender uses, by default, a camera free from any lens distortion. Lens distortion can be inserted by rendering the images using a fisheye lens, but I highly doubt this is the case at the same time it is the only lead I have.

Instructions on how to switch processing images calibrated and uncalibrated are featured in the comments.